+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Changelog for NeoFS Node
 - IR, CLI and SN verify object checksums from GET/HEAD responses (except SN proxy case) (#3406)
 - SN does not sign object GET/HEAD responses to requests with API version > v2.17 (#3406)
 - SN now verifies object header from proxy GET/HEAD responses against requested ID (#3406)
+- SN now verifies payload from proxy GET responses against in-header checksum (#3406)
 
 ### Removed